backend/requirement.txt
sqlalchemy
alembic
pydantic-settings
psycopg[binary]
SQLAlchemy>=2.0
psycopg2-binary